split typeDefs
