+++ .prettierrc
@@ -1,6 +1,6 @@
 {
     "printWidth": 120,
-    "semi": true,
+    "semi": false,
     "useTabs": true,
     "singleQuote": true,
     "trailingComma": "none",
